# doc/swagger-doc.yaml
openapi: 3.0.0
info:
  title: XRPTipBot API Specification
  description: |
   This API is a full mirror of the [XRPTipBot API](https://www.xrptipbot.com/json/feed) by [@WietseWind](https://twitter.com/WietseWind).<br>
   The API feed is updated every 30 seconds. The original API provides limited capabilities and therefore this new API was built to enhance the querying capabilities.
   
   You can use every JSON field of the feed as query parameter. Additionally these query params are supported\:<br><br>
   '<b>from_date=2019-02-01 00:03:00</b>' => to pick tips only from that date/time on<br>
   '<b>to_date=2019-02-01 00:03:00</b>' => to pick tips only until that date/time<br>
   '<b>limit=10</b>' => to limit the returned number of tips<br>

   Additionally, to speed up the requests even more, you can define a list of result fields you want to retreive in the response.<br>
   Usually the API returns the whole JSON object of the XRPTipBot-Transaction. But most likely you don`t need all the fields. So you can specify which fields you want!<br>
   The query param is called 'result_fields' and you can define the fields included in your respone, divided by comma. As Example\:<br>
   '<b>result_fields=moment,user,to,xrp</b>' -> this will return you just those 4 defined JSON properties in the response.<br>
   Try it out with this request\:<br>
   [All Tips WietseWind received](https://xrptipbot-api.siedentopf.xyz/feed?to=WietseWind&result_fields=moment,user,to,xrp)

...
  version: 0.0.2
...
paths:
...
  /aggregate/xrp:
    get:
      summary: Returns the amount of XRP for matching transactions 
      description: Returns the amount of XRP for matching transactions - depending on the filter
      tags:
        - Aggregate-API
      produces:
        - application/json
      parameters:
        - name: id
          description: id of a specific transaction
          in: query
          required: false
          schema:
            type: string
        - name: moment
          description: specific date-time you want to check the feed for
          in: query
          required: false
          schema:
            type: string
        - name: type
          description: specific type of the transaction (tip, withdraw, deposit, ...)
          in: query
          required: false
          schema:
            type: string
        - name: xrp
          description: exact value of XRP you want to check for
          in: query
          required: false
          schema:
            type: number
            format: float
        - name: network
          description: network you want to check tips for. (twitter, discord, reddit,...)
          in: query
          required: false
          schema:
            type: string
        - name: user
          description: user handle which sent a tip OR received a deposit/withdraw
          in: query
          required: false
          schema:
            type: string
        - name: to
          description: user handle which received a tip
          in: query
          required: false
          schema:
            type: string
        - name: user_id
          description: user_id which sent a tip OR received a deposit/withdraw
          in: query
          required: false
          schema:
            type: string
        - name: to_id
          description: user_id which received a tip
          in: query
          required: false
          schema:
            type: string
        - name: limit
          description: value to reduce the amount of returned tips
          in: query
          required: false
          schema:
            type: number
            fromat: integer
        - name: from_date
          description: select tips just from a specific date/time on
          in: query
          required: false
          schema:
            type: string
        - name: to_date
          description: select tips just until a specific date/time
          in: query
          required: false
          schema:
            type: string
      responses:
        200:
          description: Returns the number of matching Transactions
          content:
            application/json:
              schema:
                type: object
                properties:
                    xrp:
                      type: number
                      example: 5
        500:
          description: error
          schema:
            type: number
  /aggregate/xrp/mostReceivedFrom:
    get:
      summary: Return the users which sent the most XRP to the specified filter in a decending order
      description: Return the users which sent the most XRP to the specified filter in a decending order
      tags:
        - Aggregate-API
      produces:
        - application/json
      parameters:
        - name: id
          description: id of a specific transaction
          in: query
          required: false
          schema:
            type: string
        - name: moment
          description: specific date-time you want to check the feed for
          in: query
          required: false
          schema:
            type: string
        - name: type
          description: specific type of the transaction (tip, withdraw, deposit, ...)
          in: query
          required: false
          schema:
            type: string
        - name: xrp
          description: exact value of XRP you want to check for
          in: query
          required: false
          schema:
            type: number
            format: float
        - name: network
          description: network you want to check tips for. (twitter, discord, reddit,...)
          in: query
          required: false
          schema:
            type: string
        - name: to
          description: user handle which received a tip
          in: query
          required: false
          schema:
            type: string
        - name: to_id
          description: user_id which received a tip
          in: query
          required: false
          schema:
            type: string
        - name: limit
          description: value to reduce the amount of returned tips
          in: query
          required: false
          schema:
            type: number
            fromat: integer
        - name: from_date
          description: select tips just from a specific date/time on
          in: query
          required: false
          schema:
            type: string
        - name: to_date
          description: select tips just until a specific date/time
          in: query
          required: false
          schema:
            type: string
      responses:
        200:
          description: Returns the users with the amount of XRP received
          content:
            application/json:
              schema:
                type: object
                properties:
                    result:
                      type: array
                      items:
                        type: object
                        properties:
                            _id:
                              type: string
                              description: user_id of the XRPTipBot User
                              example: 2680350607
                            xrp:
                              type: number
                              description: amount of XRP which this user has been sent to the specified filter
                              example: 10
        500:
          description: error
          schema:
            type: number
  /aggregate/xrp/mostSentTo:
    get:
      summary: Return the users which received the most XRP from the specified filter in a decending order
      description: Return the users which received the most XRP from the specified filter in a decending order
      tags:
        - Aggregate-API
      produces:
        - application/json
      parameters:
        - name: id
          description: id of a specific transaction
          in: query
          required: false
          schema:
            type: string
        - name: moment
          description: specific date-time you want to check the feed for
          in: query
          required: false
          schema:
            type: string
        - name: type
          description: specific type of the transaction (tip, withdraw, deposit, ...)
          in: query
          required: false
          schema:
            type: string
        - name: xrp
          description: exact value of XRP you want to check for
          in: query
          required: false
          schema:
            type: number
            format: float
        - name: network
          description: network you want to check tips for. (twitter, discord, reddit,...)
          in: query
          required: false
          schema:
            type: string
        - name: user
          description: user handle which sent a tip OR received a deposit/withdraw
          in: query
          required: false
          schema:
            type: string
        - name: user_id
          description: user_id which sent a tip OR received a deposit/withdraw
          in: query
          required: false
          schema:
            type: string
        - name: limit
          description: value to reduce the amount of returned tips
          in: query
          required: false
          schema:
            type: number
            fromat: integer
        - name: from_date
          description: select tips just from a specific date/time on
          in: query
          required: false
          schema:
            type: string
        - name: to_date
          description: select tips just until a specific date/time
          in: query
          required: false
          schema:
            type: string
      responses:
        200:
          description: Returns the users and amount of XRP
          content:
            application/json:
              schema:
                type: object
                properties:
                    result:
                      type: array
                      items:
                        type: object
                        properties:
                            _id:
                              type: string
                              description: user_id of the XRPTipBot User
                              example: 2680350607
                            xrp:
                              type: number
                              description: amount of XRP which this user has been received from the specified filter
                              example: 10
        500:
          description: error
          schema:
            type: number